# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface GameLevel {
  id: number;
  title: string;
  chapter: string;
  difficulty: string;
  xp: number;
  mission: string;
  skills: string[];
  questionCount: number;
}

export type StudentInputGrade = typeof StudentInputGrade[keyof typeof StudentInputGrade];


export const StudentInputGrade = {
  NUMBER_11: '11',
  '8A': '8A',
  '8B': '8B',
} as const;

export interface StudentInput {
  /**
     * @minLength 1
     * @maxLength 80
     */
  name: string;
  grade: StudentInputGrade;
}

export type StudentProfileGrade = typeof StudentProfileGrade[keyof typeof StudentProfileGrade];


export const StudentProfileGrade = {
  NUMBER_11: '11',
  '8A': '8A',
  '8B': '8B',
} as const;

export interface StudentProfile {
  id: number;
  name: string;
  grade: StudentProfileGrade;
  xp: number;
  level: number;
  lives: number;
  completedLevels: number[];
  streak: number;
  lastSeenAt: string;
}

export interface ProgressInput {
  levelId: number;
  /** @minimum 0 */
  xpEarned: number;
  /**
     * @minimum 0
     * @maximum 30
     */
  livesRemaining: number;
  completed: boolean;
  /** @minimum 0 */
  tabSwitches?: number;
}

export interface AttemptInput {
  levelId: number;
  questionId: string;
  correct: boolean;
  /** @maxLength 12000 */
  answer?: string;
  tabSwitchDetected?: boolean;
}

export interface RunCodeInput {
  levelId: number;
  /**
     * @minLength 1
     * @maxLength 12000
     */
  code: string;
}

export type RunCodeResultStatus = typeof RunCodeResultStatus[keyof typeof RunCodeResultStatus];


export const RunCodeResultStatus = {
  passed: 'passed',
  wrong_answer: 'wrong_answer',
  syntax_error: 'syntax_error',
  runtime_error: 'runtime_error',
  timeout: 'timeout',
  service_unavailable: 'service_unavailable',
} as const;

export interface FailedTest {
  input: string;
  expected: string;
  actual: string;
}

export interface RunCodeResult {
  success: boolean;
  correct: boolean;
  output: string;
  error: string;
  sampleInput: string;
  passedTests: number;
  totalTests: number;
  message: string;
  status: RunCodeResultStatus;
  failedTest?: FailedTest;
}

export interface ChallengeSubmitInput {
  /**
     * @minLength 1
     * @maxLength 12000
     */
  code: string;
}

export interface Attempt {
  id: number;
  levelId: number;
  questionId: string;
  correct: boolean;
  createdAt: string;
}

export type StudentSummaryGrade = typeof StudentSummaryGrade[keyof typeof StudentSummaryGrade];


export const StudentSummaryGrade = {
  NUMBER_11: '11',
  '8A': '8A',
  '8B': '8B',
} as const;

export interface StudentSummary {
  id: number;
  name: string;
  grade: StudentSummaryGrade;
  xp: number;
  level: number;
  lives: number;
  completedLevels: number;
  lastSeenAt: string;
  status: string;
  tabSwitches: number;
}

export interface LeaderboardEntry {
  rank: number;
  name: string;
  xp: number;
  level: number;
  completedLevels: number;
}

export type TeacherDashboardRecentActivityItem = {
  studentName: string;
  event: string;
  levelId: number;
  createdAt: string;
};

export interface TeacherDashboard {
  totalStudents: number;
  activeToday: number;
  completedQuest: number;
  averageXp: number;
  students: StudentSummary[];
  recentActivity: TeacherDashboardRecentActivityItem[];
}