# lib/api-zod/src/generated/types/attempt.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Attempt {
  id: number;
  levelId: number;
  questionId: string;
  correct: boolean;
  createdAt: Date;
}