# lib/api-zod/src/generated/types/attemptInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface AttemptInput {
  levelId: number;
  questionId: string;
  correct: boolean;
  /** @maxLength 12000 */
  answer?: string;
  tabSwitchDetected?: boolean;
}