# lib/api-zod/src/generated/types/challengeSubmitInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ChallengeSubmitInput {
  /**
     * @minLength 1
     * @maxLength 12000
     */
  code: string;
}