# lib/api-zod/src/generated/types/failedTest.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface FailedTest {
  input: string;
  expected: string;
  actual: string;
}