# lib/api-zod/src/generated/types/gameLevel.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface GameLevel {
  id: number;
  title: string;
  chapter: string;
  difficulty: string;
  xp: number;
  mission: string;
  skills: string[];
  questionCount: number;
}