# lib/api-zod/src/generated/types/progressInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ProgressInput {
  levelId: number;
  /** @minimum 0 */
  xpEarned: number;
  /**
     * @minimum 0
     * @maximum 30
     */
  livesRemaining: number;
  completed: boolean;
  /** @minimum 0 */
  tabSwitches?: number;
}