# lib/api-zod/src/generated/types/runCodeResult.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { FailedTest } from './failedTest';
import type { RunCodeResultStatus } from './runCodeResultStatus';

export interface RunCodeResult {
  success: boolean;
  correct: boolean;
  output: string;
  error: string;
  sampleInput: string;
  passedTests: number;
  totalTests: number;
  message: string;
  status: RunCodeResultStatus;
  failedTest?: FailedTest;
}