# lib/api-zod/src/generated/types/studentInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { StudentInputGrade } from './studentInputGrade';

export interface StudentInput {
  /**
     * @minLength 1
     * @maxLength 80
     */
  name: string;
  grade: StudentInputGrade;
}