# lib/api-zod/src/generated/types/studentProfile.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { StudentProfileGrade } from './studentProfileGrade';

export interface StudentProfile {
  id: number;
  name: string;
  grade: StudentProfileGrade;
  xp: number;
  level: number;
  lives: number;
  completedLevels: number[];
  streak: number;
  lastSeenAt: Date;
}