# lib/api-zod/src/generated/types/studentSummary.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { StudentSummaryGrade } from './studentSummaryGrade';

export interface StudentSummary {
  id: number;
  name: string;
  grade: StudentSummaryGrade;
  xp: number;
  level: number;
  lives: number;
  completedLevels: number;
  lastSeenAt: Date;
  status: string;
  tabSwitches: number;
}