# lib/api-zod/src/generated/types/teacherDashboard.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { StudentSummary } from './studentSummary';
import type { TeacherDashboardRecentActivityItem } from './teacherDashboardRecentActivityItem';

export interface TeacherDashboard {
  totalStudents: number;
  activeToday: number;
  completedQuest: number;
  averageXp: number;
  students: StudentSummary[];
  recentActivity: TeacherDashboardRecentActivityItem[];
}